% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpliedVol.R
\name{ImpliedVol}
\alias{ImpliedVol}
\title{Use black76 model to compute the implied volatility of a set of European options on future contracts.
This function - ImpliedVol will return dataframe including the computed implied volatility based on original df file.}
\usage{
ImpliedVol(df)
}
\arguments{
\item{FP}{futurePrice.}

\item{K}{strike price.}

\item{type}{type of the option whether call option or put option.}

\item{Time}{time to expiry.}

\item{r}{the risk free rate, I set it as 0.02 in this package.}

\item{sigma}{the implied volatility, when it comes to initial test value, I choose the value leading Vage maximum.}

\item{ftol}{error tolerance.}
}
\value{
Dataframe including the computed implied volatility based on original df file.
}
\description{
Use black76 model to compute the implied volatility of a set of European options on future contracts.
This function - ImpliedVol will return dataframe including the computed implied volatility based on original df file.
}
