% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpliedVol.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{Plotting Implied Volatility of European Option}
\usage{
plotImpliedVol(df)
}
\arguments{
\item{df}{This dataframe should contain strike, type, optionPrice, futurePrice, and time_to_expiry of the option}
}
\description{
To understand how moneyness affect option volatility,
we plot Implied Volatility against different strike prices.
}
\details{
plotImplied function is a function that will return the implied volatility using bisection method
and plot the result

BS function is a self-written to calculate option price using Black-Scholes equation
}
