% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotImpliedVol.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{Use black76 model to compute the implied volatility of a set of European options on future contracts and plot them.
This function - plotImpliedVol will return plots based on computated implied volatility.}
\usage{
plotImpliedVol(df)
}
\arguments{
\item{FP}{futurePrice.}

\item{K}{strike price.}

\item{type}{type of the option whether call option or put option.}

\item{Time}{time to expiry.}

\item{r}{the risk free rate, I set it as 0.02 in this package.}

\item{sigma}{the implied volatility, when it comes to initial test value, I choose the value leading Vage maximum.}

\item{ftol}{error tolerance.}
}
\value{
Graph - "Implied Volatility & Time to Expiry" and "Implied Volatility & Option Price".
}
\description{
Use black76 model to compute the implied volatility of a set of European options on future contracts and plot them.
This function - plotImpliedVol will return plots based on computated implied volatility.
}
